Move analyze table logic to AnalyzeTable
